Removed secure_cookie patch and started API access
Sorry kerkouch but I have to revert your excellent secure_cookie patch because I simply do not understand how it works so I can't debug or extend it. This has prevented me working on this project for the last 6 months hoping I'd find the time to thoroughly understand how it works but I have not had the time to spare so it's easier for me to create a branch with your secure_cookie code so I can get back to working on the master branch.