Major refactor of the entire project
Reinstated some older working scripts and backed out of some Claude
based changes. Moved the old Bootstrap4 theme folder to "bootstrap4"
and renamed all the classes accoringly. Moved the new Bootstrap5 theme
to a bootstrap5 folder and similarly updated the classnames to match.

So now there are two themes that can be toggled in index.php

I also added back the old school elog(__METHOD__) to every method and
function. It can be disable by setting DGB to false. Xdebug is great
for fine grained debugging but it's not as easy and comfortable as
watching a full execution trace in the PHP log.

Next step is to go through each plugin and make sure the Create and
Update methods all use a modal popup for consistancy.