Let's start some documentation
This is with courtesy of Claude 3.5 Sonnet which I found to be quite
amazing at assisting me with code snippets and explanations.

I do not want to add comments to the code itself because they ruin
the simplicity and clean look of the code. Hence the start of these
docs to make up for the lack of code comments.